requirements.txt
Django==2.0.7
psycopg2==2.7.5
pytz==2018.5
djangorestframework==3.8.2
python-dateutil==2.7.3
